% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_3d_mesh.R
\name{build_3d_mesh}
\alias{build_3d_mesh}
\title{build_3d_mesh}
\usage{
build_3d_mesh(path2myXYZRGBtxt, search_radius, file_name)
}
\arguments{
\item{path2myXYZRGBtxt}{Path to the XYZRGB.txt input file}

\item{search_radius}{Path to the XYZRGB.txt input file}

\item{file_name}{name of the output .ply mesh file}
}
\value{
A 3D triangular mesh
}
\description{
This function reads a XYZRGB text file, requires a search radius in meters
and an output file name to save the resulting mesh.
for data format see file in package extdata folder
}
\examples{
\dontrun{path2myXYZRGBtxt<-system.file("extdata", "test.txt", package = "JFM")

file_name<- "test"

mesh3d<-build_3d_mesh(path2myXYZRGBtxt,0.5,file_name)}
}
