% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{calculate_joints}
\alias{calculate_joints}
\title{calculate_joints}
\usage{
calculate_joints(vertici_tr, indici_tri, normal_from_wild)
}
\arguments{
\item{vertici_tr}{list of facets vertexes coordinates ("vb property of mesh3d object")}

\item{indici_tri}{list of facets indexes ("it property of mesh3d object")}

\item{normal_from_wild}{matrix of data resulting from wildfire search}
}
\value{
a matrix of least square plane for each joint
}
\description{
This function calculates joint orientation with the least square method
selecting vertexes of each facet plane
}
\examples{
\dontrun{

mesh3d<-build_3d_mesh(path2myXYZRGBtxt,0.5,file_name)

normali_recalc<-Rcpp_wildfire_search(7,normals[,1:3],neighbours)

joint_list_Cpp<-calculate_joints(mesh3d,normali_recalc)}

}
