% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{calculate_joints_area}
\alias{calculate_joints_area}
\title{calculate_joints_area}
\usage{
calculate_joints_area(normal_from_wild)
}
\arguments{
\item{normal_from_wild}{matrix of data resulting from wildfire search}
}
\value{
a list of the area of each plane
}
\description{
This function calculates the area of each cluster of facets 
belonging to the same plane
}
\examples{

\dontrun{

normali_recalc<-Rcpp_wildfire_search(7,normals[,1:3],neighbours)

calculate_joints(mesh3d,normali_recalc)

}


}
