% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{compute_facets_normal}
\alias{compute_facets_normal}
\title{compute_facets_normal}
\usage{
compute_facets_normal(vertici_tr, indici_tri)
}
\arguments{
\item{vertici_tr}{list of facets vertexes coordinates ("vb property of mesh3d object")}

\item{indici_tri}{list of facets indexes ("it property of mesh3d object")}
}
\value{
matrix of the three component  of the normal vector and area of each face
}
\description{
This function returns a matrix of the three component vector of the normal of each facet.
}
\examples{
\dontrun{indici_tri<-t(mesh3d[['it']])

vertici_tr<-t(mesh3d[["vb"]])

normals<-compute_facets_normal(vertici_tr,indici_tri)}

}
