% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_triangle_area_rcpp}
\alias{compute_triangle_area_rcpp}
\title{returns the area of a mesh facet}
\arguments{
\item{tr_vertex_coords}{A 3x3 matrix of the coordinates of facet vertexes}
}
\description{
returns the area of a mesh facet
}
