% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_triangles_rcpp}
\alias{find_triangles_rcpp}
\title{returns the indexes of the neighbour facets of a target facet (nested in findNeighbourFacets function)}
\usage{
find_triangles_rcpp(indici_tr, r)
}
\arguments{
\item{indici_tr}{matrix of facets ID the "it" property of a mesh3D}

\item{r}{index of the row of the target facet}
}
\description{
returns the indexes of the neighbour facets of a target facet (nested in findNeighbourFacets function)
}
\examples{
\dontrun{indici_tri<-t(mesh3d[['it']])
row_index<-1
find_triangles_rcpp (indici_tri,row_index)}

}
