% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{findNeighbourFacets}
\alias{findNeighbourFacets}
\title{findNeighbourFacets}
\usage{
findNeighbourFacets(no_cores, indici_tri)
}
\arguments{
\item{no_cores}{number of core to use in search computation}

\item{indici_tri}{list of facets indexes ("it property of mesh3d object")}
}
\value{
a matrix of indexes of facets neighbours of target face saved on working dir
}
\description{
This function finds the IDs of each mesh facet
It requires number of cores of your pc to use and list of facets indexes
corresponding to the "it" property of mesh3d object.
}
\examples{

\dontrun{indici_tri<-t(mesh3d[['it']])
 
require("parallel")

detectCores()

no_cores <- detectCores() - 4 ### keep free some cores

neighbours<-findNeighbourFacets(no_cores,indici_tri)}
}
