% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JFM-package.R
\docType{package}
\name{JFM-package}
\alias{JFM-package}
\alias{JFM}
\title{JFM: A package for structural analysis.}
\description{
Provides functions to extract joint planes from 3D triangular mesh and makes data
available for structural analysis.
Below compute_plane_normal function description an example test over all function is done.
In your package directory you will find in extdata dir an example of point_cloud.txt file
and in test folder two R scripts a test.R file to process point_cloud.txt file and a 
JFM_workflow.R generic workflow script.
}
\author{
Riccardo Campana <riccarl@hotmail.it>
}
