% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_joint_great_circles.R
\name{plot_joint_great_circles}
\alias{plot_joint_great_circles}
\title{plot_joint_great_circles}
\usage{
plot_joint_great_circles(giac_max, file_name)
}
\arguments{
\item{giac_max}{Joint maxima poles returned from plot_joint_poles function}

\item{file_name}{Name of the output data file}
}
\value{
A plot with great circles of joint maxima saved in working dir
}
\description{
This function loads joint maxima poles, convert them to great circles and 
plot them on Schmidt stereogram.
Data are also saved in working folder.
}
\examples{

\dontrun{

poles_maxima<-plot_joint_poles(normali_recalc,joint_list_Cpp,val_area,file_name,max_pole,min_area)

azi_dip_maxima<-plot_joint_great_circles(poles_maxima, file_name)}

}
