% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maxima2mesh.R
\name{plot_maxima2mesh}
\alias{plot_maxima2mesh}
\title{plot_maxima2mesh}
\usage{
plot_maxima2mesh(mesh_tr, planes_max, normal_from_wild, tol_ang_fam)
}
\arguments{
\item{mesh_tr}{an object of type mesh3d}

\item{planes_max}{the output of plot_joint_great_circles function}

\item{normal_from_wild}{the output matrix iresulting from wildfire search}

\item{tol_ang_fam}{a tolerance angle to include joints in the same joint set color}
}
\value{
A plot with great circles of joint maxima
}
\description{
This funtion plots a coloured mesh facets as great circles plot colours
}
\examples{

\dontrun{

azi_dip_maxima<-plot_joint_great_circles(poles_maxima, file_name)

plot_maxima2mesh(mesh3d,azi_dip_maxima,normali_recalc,10)
}

}
