% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funzioni.R
\name{plotrand_col_planes}
\alias{plotrand_col_planes}
\title{plotrand_col_planes}
\usage{
plotrand_col_planes(mesh_tr, normal_from_wild)
}
\arguments{
\item{mesh_tr}{an object of type mesh3d}

\item{normal_from_wild}{the output matrix resulting from wildfire search}
}
\value{
a 3d plot of mesh with facets of the same plane
}
\description{
This function returns a 3d plot of mesh where facets of the same plane
are of same color.
}
\examples{
\dontrun{

mesh3d<-build_3d_mesh(path2myXYZRGBtxt,0.5,file_name)

normali_recalc<-Rcpp_wildfire_search(7,normals[,1:3],neighbours)

plotrand_col_planes(mesh3d,normali_recalc)}

}
