% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_crossProd}
\alias{rcpp_crossProd}
\title{returns the outer product of ab and ac}
\usage{
rcpp_crossProd(ab, ac)
}
\arguments{
\item{ab}{a 3D numeric vector}

\item{ac}{a 3D numeric vector}
}
\value{
the outer product of \code{ab} and \code{ac}
}
\description{
returns the outer product of ab and ac
}
\examples{
a1<-c(1,2,3)
a2<-c(3,4,5)
rcpp_crossProd(a1,a2)
}
