% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_wildfire_search}
\alias{Rcpp_wildfire_search}
\title{returns a matrix with the 3 components of each face normal vector; the 4th column is the ID of the plane each facet belongs to
the 5th colum the area of each facet}
\usage{
Rcpp_wildfire_search(tol_ang, list_of_normals, list_neighbours)
}
\arguments{
\item{tol_ang}{the maximum angle between facets normal belonging to the same plane}

\item{list_of_normals}{the matrix of the components of each facet normal vector}

\item{list_neighbours}{the matrix of facets ID neighbours of each target facet}
}
\value{
the IDs of same joint facets given a \code{tol_angle} between facets normal and 3Dmesh \code{list_of_normals} and \code{list_neighbours}
}
\description{
returns a matrix with the 3 components of each face normal vector; the 4th column is the ID of the plane each facet belongs to
the 5th colum the area of each facet
}
\examples{
\dontrun{neighbours<-find_neighbours_rcpp(indici_tri)
normals<-compute_facets_normal(vertici_tr,indici_tri)
tol_ang<-7
normali_recalc<-Rcpp_wildfire_search(tol_ang,normals[,1:3],neighbours)}
}
