% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_dotproduct}
\alias{rcpparma_dotproduct}
\title{returns the inner product of ab and ac}
\arguments{
\item{ab}{a 3D numeric vector}

\item{ac}{a 3D numeric vector}
}
\value{
the dot product of \code{ab} and \code{ac}
}
\description{
returns the inner product of ab and ac
}
\examples{
a1<-c(1,2,3)
a2<-c(3,4,5)
rcpparma_dotproduct(a1,a2)
}
