% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_plane_normal}
\alias{compute_plane_normal}
\title{returns the least square plane from the vertexes of facets of the same plane (nested in calculate_joints function)}
\usage{
compute_plane_normal(it_id_plane_points, vb_facets, id_fam_no_zero)
}
\arguments{
\item{it_id_plane_points}{the "it"property of mesh object binded with ID column of widfire search}

\item{vb_facets}{the vb property of mesh object (vertexes coordinates)}

\item{id_fam_no_zero}{the list of planes ID}
}
\description{
returns the least square plane from the vertexes of facets of the same plane (nested in calculate_joints function)
}
\examples{
\dontrun{ vertici_tr<-t(d_mesh[["vb"]])
indici_tri<-t(d_mesh[['it']]) 
id_fam<-normal_from_wild[!duplicated(normal_from_wild[,5]),5]
id_fam_no_zero<-id_fam[which(id_fam!=0)]
it_id_fam<-cbind(indici_tri,normal_from_wild[,5])
compute_plane_normal(it_id_fam,vertici_tr,id_fam_no_zero) }

}
