% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_neighbours_rcpp}
\alias{find_neighbours_rcpp}
\title{This function finds the IDs of each mesh facet. It requires a list of facets indexes corresponding to the "it" property of mesh3d object}
\usage{
find_neighbours_rcpp(indici_tr)
}
\arguments{
\item{indici_tr}{matrix of facets ID the "it" property of a mesh3D}
}
\description{
This function finds the IDs of each mesh facet. It requires a list of facets indexes corresponding to the "it" property of mesh3d object
}
\examples{
\dontrun{indici_tri<-t(mesh3d[['it']])
 
 neighbours<-find_neighbours_rcpp(indici_tri)}

}
