% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_triangles_rcpp}
\alias{find_triangles_rcpp}
\title{returns the row indexes of the neighbour facets of a target facet (nested in findNeighbourFacets R function)}
\usage{
find_triangles_rcpp(indici_tr, r)
}
\arguments{
\item{indici_tr}{matrix of facets ID the "it" property of a mesh3D}

\item{r}{index of the row of the target facet}
}
\value{
returns the row indexes of the neighbour facets of the facet at \code{r} row of \code{indici_tr} facet indexes matrix
}
\description{
returns the row indexes of the neighbour facets of a target facet (nested in findNeighbourFacets R function)
}
\examples{
indici_tri<-matrix(data = c(1, 2, 3 ,5, 6,
3, 2, 3, 5,7, 8 ,1),
nrow = 4,ncol = 3, byrow = TRUE)
row_index<-1
find_triangles_rcpp (indici_tri,row_index)
}
