% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{least_square_plane_rcpp}
\alias{least_square_plane_rcpp}
\title{returns the coefficients of the least square plane and the relative medium square error}
\usage{
least_square_plane_rcpp(PointsXYZ)
}
\arguments{
\item{PointsXYZ}{matrix of coordinates of point}
}
\description{
returns the coefficients of the least square plane and the relative medium square error
}
\examples{
list_xyz<-matrix(data = c(-10, -10, -15,10, -10, -5, -10, 10,5, 10, 10,15)
,nrow = 4,ncol = 3, byrow = TRUE)
least_square_plane_rcpp(list_xyz)
}
